Rename test module to avoid testing-related issues, refactor sql argument
