+++ posthog/tasks/test/test_async_migrations.py
@@ -6,7 +6,7 @@ from celery import states
 from celery.result import AsyncResult
 from constance.test import override_config
 
-from posthog.async_migrations.examples.test import Migration
+from posthog.async_migrations.examples.test_migration import Migration
 from posthog.async_migrations.runner import run_async_migration_next_op, run_async_migration_operations
 from posthog.async_migrations.test.util import create_async_migration
 from posthog.models.async_migration import AsyncMigration, MigrationStatus
